[Bronze III] Title: 팰린드롬인지 확인하기, Time: 0 ms, Memory: 1116 KB -BaekjoonHub

+++ 백준/Bronze/10988. 팰린드롬인지 확인하기/팰린드롬인지 확인하기.c
@@ -7,7 +7,7 @@ int main(void) {
     scanf("%s", s);
 
     int i = 0;
-    int j = (int) strlen(s) - 1;
+    int j = strlen(s) - 1;
 
     while (i < j) {
         if (s[i] != s[j]) {
